Refactor app into modular components with separate tab rendering
